[BL] Move IntegerSequence into Functional.
